requirements.txt
appdirs==1.4.3
Django==1.11
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-imagekit==4.0
django-recaptcha==1.3.0
Markdown==2.6.8
mistune==0.7.4
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.1.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
